% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRbs.R
\docType{package}
\name{aRbs}
\alias{aRbs}
\title{aRbs: A package for finding arbitrage opportunities}
\description{
This package finds arbitrage opportunities from www.oddschecker.com and
provides useful information so that the user can exploit this arbs in the
most effective way.
}
\author{
Andrew Little \email{andrewlittlebristol@gmail.com}
}
