% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_links.R
\name{clean_links}
\alias{clean_links}
\title{Clean links given by the \code{scrape_links()} function.}
\usage{
clean_links(links)
}
\arguments{
\item{links}{A \code{data.frame} of links with a column "url" containing
www.oddschecker.com URLs, usually the output of the \code{scrape_links()} function.}
}
\description{
Filter links to only those ending in "winner" and clean for use in the
\code{get_arb_single()} function.
}
\examples{
links <- scrape_links("https://www.oddschecker.com/football")
clean_links(links)
}
