% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddschecker2.R
\name{oddschecker2}
\alias{oddschecker2}
\title{Find the odds of each outcome of an event from www.oddschecker.com}
\usage{
oddschecker2(event, in_play = TRUE)
}
\arguments{
\item{event}{A www.oddschecker.com event path, given as a string. This is essentially
an event URL with the "www.oddschecker.com/" base removed.}

\item{in_play}{Logical. Should in-play arbitrage opportunities (arbs) also be returned?
If \code{FALSE} and event is in-play, then a NULL \code{data.frame} is returned.}
}
\description{
Find the odds for each of the listed outcomes of an event on www.oddschecker.com, for
each of the listed bookies. This function is essentially a re-exported fix of the
function \code{oddschecker} from the {gambleR} package.
}
