% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.arb.R
\name{print.arb}
\alias{print.arb}
\title{Print method for class arb}
\usage{
\method{print}{arb}(arb, ...)
}
\arguments{
\item{arb}{An object of class \code{arb} - usually the output of a call to
\code{get_arbs()}}

\item{...}{Arguments passed to \code{base::print.default()}.}
}
\description{
An S3 method for printing objects of class \code{arb}. This will print the
\code{best_choice} elements from
}
\details{
Prints the the best choices of arbitrage opportunities.
}
