% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_form.R
\name{remove_form}
\alias{remove_form}
\title{Fix appended form bug}
\usage{
remove_form(string)
}
\arguments{
\item{string}{Any string. If ending in a 5 character combination from the set
{"w", "d", "l"} then this will be removed. Otherwise the string will be kept
as is.}
}
\description{
This function removes any appended form from a string
}
\details{
Sometimes when generating outcome names, the form of a team can be appended
to the name. For example, if looking at the odds for St Johnstone FC vs
Dundee FC, St Johnstone may be written as St Johnstonewwldd
if their previous five games resulted Win, Win, Loss, Draw, Draw. This can
be removed with \code{remove_form()}.
}
\keyword{internal}
