% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_links.R
\name{scrape_links}
\alias{scrape_links}
\title{Scrape all subdomain URLs of a domain}
\usage{
scrape_links(url)
}
\arguments{
\item{url}{A head domain URl, given as a string.}
}
\value{
A \code{data.frame}.
}
\description{
Scrape all subdomain URLs of a domain
}
\details{
Used in the call to \code{get_arbs()} to find subdomains that
 may offer events, from which we can find an arb.
}
\examples{
scrape_links("https://www.oddschecker.com/football")
}
