% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddschecker2.R
\name{oddschecker2}
\alias{oddschecker2}
\title{Find the odds of each outcome of an event from www.oddschecker.com}
\usage{
oddschecker2(event)
}
\arguments{
\item{event}{A www.oddschecker.com event path, given as a string. This is essentially
an event URL with the "www.oddschecker.com/" base removed.}
}
\description{
Find the odds for each of the listed outcomes of an event on www.oddschecker.com, for
each of the listed bookies. This function is essentially a re-exported fix of the
function \code{oddschecker} from the {gambleR} package.
}
